split doctors appointment and doctor endpoints
